Updated menu.py | Removed teams.py
Updated menu.py with better visuals and renamed to collector.py. Removed the need for the teams.py(Team class). Data is now being stored in a list as long as the program is running. Need to figure out how to store it in json format. add_team, add_player, remove_player, & update_player functions all added to collector.py.